// artisanat-frontend/src/app/core/enums/specialty.enum.ts
export enum Specialty {
  WEAVING = 'WEAVING',
  EMBROIDERY = 'EMBROIDERY',
  CARPET_MAKING = 'CARPET_MAKING',
  LEATHERWORK = 'LEATHERWORK',
  SHOE_MAKING = 'SHOE_MAKING',
  WOOD_CARVING = 'WOOD_CARVING',
  FURNITURE_MAKING = 'FURNITURE_MAKING',
  MARQUETRY = 'MARQUETRY',
  JEWELRY_MAKING = 'JEWELRY_MAKING',
  METALWORK = 'METALWORK',
  POTTERY = 'POTTERY',
  STONEMASONRY = 'STONEMASONRY',
  GLASSBLOWING = 'GLASSBLOWING',
  TAPESTRY = 'TAPESTRY',
  MOSAIC_MAKING = 'MOSAIC_MAKING'
}